:recycle: 공동체 생성 react-query 옵션 위치 변경 (#345)
